requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
beanie==1.22.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.23.5
httpx==0.26.0
swagger-ui-bundle==1.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0